# RUST-2239 Fix performance of bson 3.0