--- a/src/macros.rs
+++ b/src/macros.rs
@@ -439,6 +439,7 @@ macro_rules! rawdoc {
 ///
 /// This macro generates a `SerializeAs`/`DeserializeAs` implementation for a given type,
 /// with optional struct-level attributes like `#[derive(...)]` or `/// doc comments`.
+#[allow(unused_macros)]
 macro_rules! serde_conv_doc {
     ($(#[$meta:meta])* $vis:vis $m:ident, $t:ty, $ser:expr, $de:expr) => {
         #[allow(non_camel_case_types)]
@@ -490,4 +491,5 @@ macro_rules! serde_conv_doc {
     };
 }
 
+#[allow(unused_imports)]
 pub(crate) use serde_conv_doc;