… src/raw/document_buf/raw_writer.rs